+++ tests/testthat/test-defunct.r
@@ -1,0 +1,7 @@
+test_that("generate informative errors", {
+  expect_snapshot(error = TRUE, {
+    id()
+    failwith()
+    funs()
+  })
+})